Support highlighting file numbers (#919)
This PR adds support for highlighting of file numbers in case law search
results.

In order to simplify the logic, the textMatch key conversion (from
snake_case to camelCase, dropping `.text` suffix where used) is now done
in all cases, not just in the AllDocumentsSearchController.

Test cases have been added. In the edge case that a file number is input
without any whitespace or special characters, the highlighting is not
expected.