Change ObjectStorage method naming, exceptions (#924)
This PR changes the naming of ObjectStorage methods to be more in line
with standard object storage terms, like using "object" instead of
"file" and "key" instead of "path".
It also changes the RetryableObjectStoreException to a more generic
ObjectStoreServiceException, with a custom NoSuchKeyException to
differentiate such cases.

The `get` method is removed to enforce more specific exception handling.
Up to this point, any kind of exception thrown in this method would
cause clients to get a `404` error code.