Use correct Docker image tag (#922)
Sometimes, E2E pipelines may not use the most recent frontend or backend
image, if that image has been built in the previous step. This PR
changes this to use the github.sha value, if applicable.
The frontend E2E pipeline will still use the PR-branch or main backend
image, and vice versa. This should be acceptable for most cases.