RISDEV-7621 add different searches to browser history (#930)
This PR changes `router.replace` to `router.push` when propagating Pinia
state changes to the browser History API, and adds logic to enable the
store to update according to the URL.

In order to prevent cycles where the store updates the route and the
route change triggers a store update, logic is added to ensure that the
store ignores route changes triggered by it.

Test cases have been added, including for PostHog reporting.