Replace Axios search by useFetch (#936)
This PR replaces the Axios searchService through useAsyncData with a
more Nuxt-standard `useFetch` implementation.
The advantage is that useFetch handles passing client credentials to the
server in SSR mode. It is also more concise.

The motivation behind this PR is to enable search result rendering on
the server side as well. Previously, SSR would fail because the client
cookie value wasn't being passed to the backend when running in
authenticated mode.