RISDEV-7808 Search for common article names by adding extra field (#945)
This PR enables searching for articles by number and norm abbreviation,
e.g., "§ 3 BGB".
To this end, a new helper field is added to nested list "articles" in
OpenSearch.

The nested query ScoreMode is updated from "none" to allow for nested
content to influence ranking.

A data test is added to check for matches. Note that it only checks for
"§ 3 BGB", not "BGB 3". The inverse condition could be added, if needed.