Fix CSP errors concerning Sentry, SVG (#955)
The Sentry module and an SVG image were blocked by CSP, due to the fact
that
- the Sentry Nuxt module is not considered by the Nuxt-security module
- the SVG image was blocked (presumably) due to disallowed inline styles

This PR moves the Sentry configuration to a useScript directive, to be
included by the main script under a `strict-dynamic` policy.
It also updates the development configuration to allow for hot reloading
using a worker again.