RISDEV-7560 Remove talisman and introduce github secrets check (#932)
This PR replaces the use of **Talisman** with **GitHub Secrets Push
Prevention** to simplify the development workflow and enhance security.
The decision aligns with the repository's transition to being public,
leveraging GitHub's built-in features for secret detection.

The following has changed as part of this PR
1. The pipeline for secrets check using talisman has been removed since
it's prevented originally from being pushed.
2. The talisman configuration file has been removed.
3. The details of the decision and consequences have been documented in
an ADR.
4. The lefthook has been simplified and updated accordingly.

---------

Co-authored-by: Carl Gödecken <carl.goedecken@digitalservice.bund.de>