RISDEV-8006 Page for historic versions of a norm (#973)
This PR makes navigation between different versions of a norm more
convenient.

The main feature is a new breadcrumb element that links back to the
"most relevant" version of a norm. It directs to a Nuxt route that
performs a redirect,
- to the currently active version if available, or
- to a future or past version otherwise.
The last breadcrumb item shows the validity range of the currently open
version, unless the open version is in force.

A few end-to-end cases and unit tests were added.

---------

Co-authored-by: Haytham Mones <haythameyd@gmail.com>