RISDEV-8592: Make imports more consistent in the frontend (#1045)
- add a rule to make sure imports are always sorted and kept in a
consistent order
- add a rule which disallows the usage of `"@/..."` and enforces to use
`"~/..."` instead. I know this is quite a hard restriction but as ~ and
@ both point to the source directory, I think it adds consistency to
only use one of them. Ideally most nuxt related imports are auto imports
and are not even explicitly needed. The comments
[here](https://github.com/nuxt/nuxt/pull/24574) seem like there is a
tendency to preferr "\~" over "@" in the nuxt team - thats why I decided
for "~". See [this commit
](https://github.com/digitalservicebund/ris-search/commit/b4b4cc02593ec1705c0dc8d79614e0dbe1ba31ad)
for the actual I rules added. The second commit is just adapting the
imports to conform to the new rules.