RISDEV-8763 Create sitemap for portal static pages (#1083)
This PR implements a sitemap.xml for the static pages of the portal It
utilizes the SEO NuxtJS modules and more specifically the sitemap module
in order to build the sitemap dynamically.

Dynamic norms and caselaw pages are leftout of the scope of this PR as
they will be implemented separately and would be done at indexing time.