RISDEV-8770 Implement sitemap files for caselaw
This PR adds sitemap files for Caselaw pages.
This is entirely different from the ECLI crawler approach.

The Caselaw sitemap files are added in a separate folder next to the norms files on the portal bucket
They are stored under `sitemaps/caselaw` key on the bucket.

Additionally, the issue of using JAXB for generating the sitemap XML files has been addressed in this PR
given the repeated use of the XML generation and to avoid code duplication.