RISDEV-8857 (#1111)
Switch from cross fields to pre tokenization
* Refactor UniversalDocumentQueryBuilder into 3 parts. A shared part, a
norm part and a case law part.
* Refactor how quoted and unquoted sections are parsed from searchTerm.
Now if there is an unmatched ", the last one is removed before parsing.
This simplifies the parsing logic.
* The exact way that PortalQueryBuilder keeps it's state until the query
is built is a potential topic for a future refactor.
* The code that was moved to CaseLawQueryBuilder and NormQueryBuilder
look to have some bugs and are a topic for future work.