RISDEV-8770 Implement sitemap files for caselaw (#1107)
This PR adds sitemap files for Caselaw pages and implement the code
changes requested by the dev team during the latest dev exchange on the
topic of sitemaps.

Note: This PR scope is completely different from the ECLI crawler
approach.

- The Caselaw sitemap files are added in a separate folder next to the
norms files on the portal bucket They are stored under
`sitemaps/caselaw` key on the bucket.
- The path for the norms sitemaps remain unchanged.
- JAXB is used for generating the sitemap XML files has been addressed
in this PR given the repeated use of the XML generation and to avoid
code duplication.
- The sitemap generation process is currently a separate job that can be
triggered apart from the indexing.