RISDEV-8402 (#1159)
Simple search | Show most relevant expression (backend)
* Added EliFile and refactored WorkEli, ExpressionEli and
ManifestationEli classes.
* Norm attachments can now be an instance of EliFile, but not
ManifestationEli. The code using the classes was adjusted to reflect
this and now the names make more sense.
* Remove subtype from WorkEli because it was removed in ldml.de 1.8
* Remove subtype from ExpressionEli because it was removed in ldml.de
1.8
* Deleted dead code (and it's tests) that was only used in tests and not
in the actual code
* Refactored IndexNormsService to group by norm work so that if any file
related to a work is touched, then the whole work will be reindexed.
This refactor is necessary for the next part of the user story.
Calculation of the relevance window requires all expressions in the
entire work.