feat: data field picker (#1161)
RISDEV-8432

This PR implements the data field picker for the upcoming advanced
search page. This includes replacing the content of the current advanced
search page, however I haven't done a thorough cleanup of the old code
yet. This will be done once the full page is coming together. It also
touches some things that are only tangentially related to the component
(not super clean I know, sorry):

- adds `@testing-library/user-event` and `@testing-library/dom` (see
Slack for context)
- enables the PrimeVue plugin globally for tests since we need that
basically everywhere
- pins the version of nwsapi (see
[9d4478e](https://github.com/digitalservicebund/ris-search/pull/1161/commits/9d4478e05a6c86cb1af3fee0767a37b4342265c6)
for context)
- fixes some unrelated code quality issues that showed up in SonarCloud
for some reason

For the person reviewing this, I would be interested particularly in
feedback regarding

- if the code follows the coding practices and conventions in the rest
of the code base as I'm not super familiar with it yet
- accessibility of the new component