Literature backend API and service (#1186)
First of all sorry for the big single commit. Was my first work on the
portal backend and I learnt how it is built by coding.

Many additions and some changes in already existing code. Just
mentioning the most important changes for clarity (and stating my
reasons):

- Added field `recordingDate` to the literature model. Firstly I needed
a date field for the default sorting behaviour (for which the shared
DATUM field is used, which points to a date field in the respective
models), and secondly I think this field should be part of the model per
se, as it seems to convey important information
- Accordingly adapted the mappers to transmit this new field, including
the extraction of the field itself from the XML for which I extended the
`FrbrWork` class for literature and added the `FrbrDate` class.
- Added missing mapping definitions for literature for opensearch


And here is the list of changes done but fell into the categories
not-sure-if-it-is-right OR it-was-not-specified-and-decided-myself:
- Static method `getLiteratureHighlighter` in the `RisHighlightBuilder`
pointing to `Literature.Fields.MAIN_TITLE`
- I chose the following fields of Literature to be directly searchable
in `LiteratureSeachParams`: `documentNumber`, `yearOfPublication`,
`documentType`, `author`, `collaborator`
- The `searchAndFilterLiterature` method of the `LiteratureService` is
not being unit tested but directly integration-tested (same as in
caselaw and norms)
- No unit tests for `LiteratureQueryBuilder` and
`LiteratureSearchParams` but I don't think they are needed, specially
for the latter
- What fields should be included in the `LiteratureSearchSchema`? For
now I just copied over all the fields from the model itself
(`Literature`) but is not clear to me what fields should NOT be return
(probably `outline` and `shortReport`, which I have anyway excluded from
the search results because of long text?)
- Is the `LiteratureSortParam` ok? As mentioned above I added the DATUM
mapping for opensearch pointing to the new field `recordingDate`
- Default encoding values returned of HTML, XML but also ZIP. But there
is no ZIP endpoint cause is not needed right? But then the possible
encodings should only be HTML and XML? Change needed?
- I didn't modify the literature_mappings.json file, apart from adding
the `recordingDate` field and the default shared `DATUM` one. Meaning
without having keywords for literature the searching/filtering by array
fields is done in the `LiteratureQueryBuilder` using `matchQuery`
instead of `termQuery`, expecting an exact match as far as I understand.
Also the queries are built like OR when multiple values in the same
fields and with AND when different fields. Is this the expected way?
Felt to me more natural and permissive and I think for caselaw filtering
for the `types` field, is done also with OR