RISDEV-8944: Add literature detail page (#1219)
### This PR:
- adds a first simple version of a literature detail page implementation
- displays the title and a metadata (NOTE: the final data formatting for
the metadata section is still not determined)
- shows only the html in the text tab

### Still missing:
- actions menu with print, xml, pdf, buttons...
- details tab content
- finalizing how the metadata is displayed (singel vs plural labels,
abbreviations vs. full spelling,...)

### For the reviewer:
- I noticed there are no unit tests for the `RisBreadcrumb` component.
Tried to add one but failed. Would love some hints how to test it
properly. The implementation of the component also seems overcomplecated
to me (on first sight) so maybe its even worth a refactor. But didn't
want to start that in this PR.
- does it make sense that I move the title into a separate
RisDocumentTitle component?
- are the tests I added okay? Or should they be even more granular?