chore: rebuild lockfile
With the previous lockfile, the build fails because of some parsing
error. I noticed that the build is using a Vite beta, possibly related
to some outdated entries in the lockfile. Regenerating the lockfile
causes the build to use a more recent, non-beta version of Vite, where
the build error doesn't occur.

RISDEV-0000