Include literature in the simple search results (#1228)
Some key changes:

In the backend:
- Renamed in schemas the `alternativeTitle` to `documentaryTitle`,
simple reason, the XML model was named already like that and the index
has the same name for the field.
- Extended the literature namespace to also include both
RIS_SELBSTSTAENDIG_NS and RIS_UNSELBSTSTAENDIG_NS (this change was done
locally a long time ago, before @UrsKahmann updated that file too)
- Removed the _Kurzreferat_ (shortReport) from the excluded fields,
since it is needed for the result
- Added _dokumentarischer Titel_ and _Kurzreferat_ to the list of
highlightable fields

In the frontend:
- Mains stuff is in the `LiteratureSearchResult`, here I did some
slicing for the displaying the highlighted matches within the short
report but it feels quite dirty. Some context. The short report should
be not displayed completely but only 3 lines, for that I used at the
beginning just the tailwind class `line-clamp-3`, what worked perfectly.
But this was only useful when there were no text matches within the
short report. Since the short report is usually more than 3 lines, it
needs to be truncated but if the match should be, at least in my
opinion, as central as possible to the truncated 3 lines. The function
`getShortReportSnippet` does this (not in a perfect way, specially
because the counting of characters set to 270 is just an approx). If
there are no matches in the short report, the tailwind class
line-clamp-3 is used
- I also refactored the `getUrl` function in `SimpleSearch.logic`, i
find it his way much more readable
- Corrected also one small mistake in the breadcrumb route for
literature (was pointing to caselaw)