Add sentry minimum log level to prod
And also remove the not-needed bean in the SentryConfig

Why?
The BeforeSendCallback only affects events that actually reach the Sentry SDK, meaning anything that is about to be sent to Sentry. But when using the sentry logging integration with

sentry:
  logging:
    minimum-event-level: error

this will only forward to Sentry error/fatal logs, meaning the Sentry SDK wouldn't need to do an extra filter.

So only if you use in the backend code things like
Sentry.captureMessage("Something failed", SentryLevel.ERROR);
Then that Bean for extra filtering would be needed