Risdev 9188 link to first future version (#1239)
This PR changes the behavior of the message box in case a future version
of a Fassung exists.
Instead of referencing the latest future version of a Fassung, we want
to display the next future Fassung.

This should also solve the following issue:

- in case a version with an undefined temporal coverage exists, we
currently don't display the next future version, even though one exists