Mandatory fields literature (#1242)
What this PR contains. Three main things:

1. Throw a `ValidationException` for all literature fields that are
listed as mandatory in
[here](https://digitalservicebund.atlassian.net/wiki/spaces/VER/pages/1986068481/Rubriken+bersicht+-+Literatur).
Only document number, document types and publication years
2. Wrap all exceptions being thrown in the mapper class into the wrapper
`OpenSearchMapperException`, as it is done for caselaw documents. Then
log.error it in the literature index service (agreement of today's
discussion about how to raise error without throwing exceptions)
3. Add a utils class for loading all types of documents in tests -->
`LoadXmlUtils`. It can not only be used for literature but also for
caselaw and norms

And did also a big of refactoring in tests for loading the real XML
content as string and not fake it with a template string anymore. Also
combine many of the single tests of the mapper into one