limit feedback requests, cache with caffeine (#1245)
This PR introduces 2 Rate limit interceptors:

- rate-limiter for the feedback endpoint
- rate-limiter for default behaviour


Behavior of both is configured in the applications.yml.
The rate-limiting allows a specified number of requests for a specified
amount of time.
Limiting is based on the remote address of an incoming request.

For now, state of the rate-limiting is stored in memory.