RISDEV-8944: Literature details tab (#1274)
**This PR:**
- implements the literature details tab
- improves the description list (adding labels)
- adds tests 

**I went a bit out of scope (sorry) with**:
- in order to reuse the function for formatting arrays, I wanted to
remove the default placeholder '-' value from it
- I think the placeholder should be defined by the component anyways and
not the formatting function
- unifies the placeholder that is used by the metadata fields (see [this
slack thread for
context](https://digitalservicebund.slack.com/archives/C05J19X1E6L/p1761562120111459))
- adapts tests (for the caselaw page the tests are a mix of the vue
testing libraray and the vue test utils, not ideal but changing all test
in the file was out of scope, furthermore I am not sure if we will
actually keep this testfile around)

**Sonar Cloud Issue:**
- besides the issue of code duplication in tests it raised the problem
that a `<dt>` tag is not surrounded by a `<dl>`. That is because the tow
parts (List and items) are defined as two separate components.
Refactoring that would blow up this PR too much I think, thus I would
like to get this PR merged first and then handle that issue. Let me know
if you see any problems with that :)