RISDEV-10028: Norm relevance window (#1372)
## This PR:
- changes the simple search to use the `mostRelevantOn` parameter when
searching for norms insetad of the `temporalCoverage`
- refactors the function to convert the parameters and adds tests
- adds e2e test to check that the correct norms are returned, correct in
the sense that the mostRelevant one for the current date is returned
- adds an e2e test that makes sure the all documents search always
returns documents of all document types (when no search parameters are
given)

## Questions/Notes for the reviewer:
- I set fixed times for some of the playwright tests to test the
validity behaviour of norms, this leads to hydration issues when the
tests are run. I have no better solution but I am unsure as well whether
this is acceptable or not. Would love to hear your thoughts on it.