RISDEV-0000 improve baseline for frontend unit tests (#1392)
This PR adds unit tests to frontend code that didn't previously have
any. Some notes:

- ⚠️ All tests have been generated with AI in a "quantity over quality"
approach. Obviously many could be improved or more extensive if someone
more familiar with the project wrote them by hand. However the goal is
not for them to be perfect but to raise the baseline of behavior and
components that are covered by automated testing while keeping effort
minimal.

- Tests should meet the following minimum requirements:
  - A reasonable default configuration of the component renders without
errors
  - Props are reflected correctly in the UI
  - Slots are rendered
  - Defined events are emitted
  - Models work as expected

- Many of the tests were probably missing because the components are
trivial, IMO that is not a reason to at least test that it renders, so I
added those tests, too

- While at it, I updated the code coverage exclusions, to reflect our
team agreements and removed paths that don't exist anyway

- Also note that I did not generate tests for things that should not
exist in the first place 😅 this includes the icon components and
`httpService.ts`

- **Future improvements:**
  - Align naming conventions - remove all `.unit` extensions
  - Remove existing integration tests in `src/pages` in favor of
Playwright tests, either refactor or delete depending on existing E2E
tests
  - Remove stubbing from unit tests where possible, IMO this needlessly
complicates the tests