RISDEV-9102: Show VV search results in all document search (#1425)
### This PR:
- use unified search result header for all search result kinds
    - this should also fix the icon alignment with the headline 
- add administrative directive search result component and tests
- this does not yet add the VV filter options and VV specific search to
the simple search, this will follow in a separate PR

### Questions to the Reviewer
- I don't really like the `isMarkup` flag on the SearchResultHeaderItem
but didn't find a better way. Other options would be to render them all
as markup.. Idk. Any ideas to improve that?
- I am not super happy with some of the tests.
- for the SearchResultHeader I would have liked to assert the whole
header as one string instead of each element by itself
- for testing that markup is rendered correctly I sticked to how it was
done in other test, though I am not sure if that is the cleanest
appraoch