RISDEV-9032: Add metadata to administrative directive page (#1419)
### This PR
- refactors the metadata section and unifies its usage for caselaw,
literature and administrative directives
- use a description list to represent the label-value pairs instead of
divs
- I didn't change the norms metadata section yet as this is a little
more complex and the PR is already large

### Questions to the Reviewer:
- I struggled really hard with the description list, specifically:
- this list itself does not get a role (at least none that is recognized
by playwright) which makes it hard to target the list elements directly
in tests. I now get all the `term` and `definition` elements but this
will break if there are more `<dl>` in the page/component
- I noticed that for the `DetailsProperty` component (which also useses
dt and dd) we (I did it 😓) use aria labled by which I think is not the
correct way (at least voicover seems to read out the terms twice this
way :/). So I put it on my todo list to change that there as well
- not having the aria-labledby makes it kind of ugly to test the
description list because the terms and definition are now only connected
by their order
- overall I spend a lot of time figuring out what the correct way to use
and test the description list is, but I am still not sure if the
solution is okay now :D