RISDEV-8632 Fix eingangsformel anchor in page (#1458)
This PR fixes the issue with the eingangsformel and any other element in
the norms' view page that has an anchor element that has been encoded by
changing umlauts to regular characters. The links in the table of
content and norm search results have been adjusted to point to the
correct id of the anchored header of the section in the norm view page.