in case raw ids cant be found, try encoded as fallback (#1513)
The article ids we potentially receive to create an html represenation
from aren't necessarily encoded in the same way.
It is possible we get:
- special characters -> we need to encode them to find their actual ids
in ldml (special characters are encoded in ldml)
- umlaute -> we shouldn't encode to find the appropriate id in ldml
(umlaute are unencoded in ldml)

To handle this until we have unified behaviour in ldml, we decided to
first try to parse the id raw, and in case nothing is found try to find
its UTF-8 encoded presentation as a fallback.
This avoids having to introduce several permutations of possible edge
cases we might encounter.

We also moved the logic from the controller to the xslt transform
service, to enforce similiar behaviour across usages.