RISDEV-10337 Fix the table of content link to article within norm view
This PR is a hotfix to the following issue:
While our current uri-encoded eIDs work for cross-page linking, they break same-page anchor scrolling. Browsers automatically decode those parts in the url (e.g., #art-z%c2%a7 becomes #art-z§), but since our HTML IDs are literally stored in the encoded format, the browser fails to find a match in the DOM.
Proposed Short-term Fix: I can implement a click-interceptor in the TOC (table of content) component. It will catch the anchor click, manually match the encoded ID, and trigger scrollIntoView.
- Pro: Immediate fix for the current UI breaking.
- Con: Relies on JS; doesn't solve the underlying data inconsistency and would fail if noJS is in place.

This fix is agreed to be a temporary fix and a more permanent fix is to be thought of and implemented later.