RISDEV-10337-norms-toc-anchors (#1517)
This PR replaces the current workaround to get the table of contents for
norms to work with article ids that contain special characters e.g.
#art-z%c2%a7%c2%a7%2015%20bis%2018. This approach doesn't work without
javascript.

This PR replaces the NuxtLink for the navigation of the table of
contents items with plain anchors, to prevent normalization of the ids
(default behaviour of vue router). The <a> will work with the raw ids,
without normalization. This should avoid the issue of ids not being
found due to decoding of the hash